test(select): migrate legacy standalone
